Refactor directories for pytest tests
