+++ pytest.ini
@@ -5,4 +5,4 @@ filterwarnings =
 env_files =
     test.env
 testpaths =
-    integration_tests/tests
+    tests/functional
